move legacy helpers and converters
